requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.7
deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
frozenlist==1.7.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
kombu==5.5.4
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.27
ldap3==2.9.1
line-bot-sdk==3.19.0
loguru==0.7.3
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
minio==7.2.16
multidict==6.6.4
openai==1.107.1
orjson==3.11.3
packaging==25.0
passlib==1.7.4
pillow==11.3.0
playwright==1.55.0
pluggy==1.6.0
prompt-toolkit==3.0.52
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyee==13.0.0
pygments==2.19.2
pyjwt==2.10.1
pymupdf==1.26.4
pyodbc==5.2.0
pytesseract==0.3.13
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pyyaml==6.0.2
redis==5.2.1
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.37.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.47.3
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.11.0
tqdm==4.67.1
typer==0.17.4
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
win32-setctime==1.2.0
wrapt==1.17.3
yarl==1.20.1
zstandard==0.24.0
